+++ grpc-slaughterhouse/src/main/java/slaughterhouse/grpc/AnimalsIdList.java
@@ -85,13 +85,13 @@ private static final long serialVersionUID = 0L;
   }
   public static final com.google.protobuf.Descriptors.Descriptor
       getDescriptor() {
-    return slaughterhouse.grpc.RecallService.internal_static_slaughterhouse_grpc_AnimalsIdList_descriptor;
+    return slaughterhouse.grpc.RecallServiceOuterClass.internal_static_slaughterhouse_grpc_AnimalsIdList_descriptor;
   }
 
   @java.lang.Override
   protected com.google.protobuf.GeneratedMessageV3.FieldAccessorTable
       internalGetFieldAccessorTable() {
-    return slaughterhouse.grpc.RecallService.internal_static_slaughterhouse_grpc_AnimalsIdList_fieldAccessorTable
+    return slaughterhouse.grpc.RecallServiceOuterClass.internal_static_slaughterhouse_grpc_AnimalsIdList_fieldAccessorTable
         .ensureFieldAccessorsInitialized(
             slaughterhouse.grpc.AnimalsIdList.class, slaughterhouse.grpc.AnimalsIdList.Builder.class);
   }
@@ -302,13 +302,13 @@ private static final long serialVersionUID = 0L;
       slaughterhouse.grpc.AnimalsIdListOrBuilder {
     public static final com.google.protobuf.Descriptors.Descriptor
         getDescriptor() {
-      return slaughterhouse.grpc.RecallService.internal_static_slaughterhouse_grpc_AnimalsIdList_descriptor;
+      return slaughterhouse.grpc.RecallServiceOuterClass.internal_static_slaughterhouse_grpc_AnimalsIdList_descriptor;
     }
 
     @java.lang.Override
     protected com.google.protobuf.GeneratedMessageV3.FieldAccessorTable
         internalGetFieldAccessorTable() {
-      return slaughterhouse.grpc.RecallService.internal_static_slaughterhouse_grpc_AnimalsIdList_fieldAccessorTable
+      return slaughterhouse.grpc.RecallServiceOuterClass.internal_static_slaughterhouse_grpc_AnimalsIdList_fieldAccessorTable
           .ensureFieldAccessorsInitialized(
               slaughterhouse.grpc.AnimalsIdList.class, slaughterhouse.grpc.AnimalsIdList.Builder.class);
     }
@@ -344,7 +344,7 @@ private static final long serialVersionUID = 0L;
     @java.lang.Override
     public com.google.protobuf.Descriptors.Descriptor
         getDescriptorForType() {
-      return slaughterhouse.grpc.RecallService.internal_static_slaughterhouse_grpc_AnimalsIdList_descriptor;
+      return slaughterhouse.grpc.RecallServiceOuterClass.internal_static_slaughterhouse_grpc_AnimalsIdList_descriptor;
     }
 
     @java.lang.Override
