+++ grpc-slaughterhouse/src/main/java/slaughterhouse/grpc/RecallServiceOuterClass.java
@@ -3,8 +3,8 @@
 
 package slaughterhouse.grpc;
 
-public final class RecallService {
-  private RecallService() {}
+public final class RecallServiceOuterClass {
+  private RecallServiceOuterClass() {}
   public static void registerAllExtensions(
       com.google.protobuf.ExtensionRegistryLite registry) {
   }
@@ -48,13 +48,12 @@ public final class RecallService {
       "\n\n\002id\030\001 \001(\005\"<\n\rAnimalsIdList\022+\n\004list\030\001 \003" +
       "(\0132\035.slaughterhouse.grpc.AnimalId\"=\n\rPro" +
       "ductIdList\022,\n\004list\030\001 \003(\0132\036.slaughterhous" +
-      "e.grpc.ProductId2m\n\022ProductsWithAnimal\022W" +
-      "\n\022productsWithAnimal\022\035.slaughterhouse.gr" +
-      "pc.AnimalId\032\".slaughterhouse.grpc.Produc" +
-      "tIdList2j\n\020AnimalsInProduct\022V\n\020animalsIn" +
-      "Product\022\036.slaughterhouse.grpc.ProductId\032" +
-      "\".slaughterhouse.grpc.AnimalsIdListB\002P\001b" +
-      "\006proto3"
+      "e.grpc.ProductId2\300\001\n\rRecallService\022W\n\022pr" +
+      "oductsWithAnimal\022\035.slaughterhouse.grpc.A" +
+      "nimalId\032\".slaughterhouse.grpc.ProductIdL" +
+      "ist\022V\n\020animalsInProduct\022\036.slaughterhouse" +
+      ".grpc.ProductId\032\".slaughterhouse.grpc.An" +
+      "imalsIdListB\002P\001b\006proto3"
     };
     descriptor = com.google.protobuf.Descriptors.FileDescriptor
       .internalBuildGeneratedFileFrom(descriptorData,
